+++ src/MaybeResults/MaybeExtensions.cs
@@ -2,109 +2,114 @@
 
 public static class MaybeExtensions
 {
-    public static IMaybe<TOut> Map<TOut>(this IMaybe result, Func<TOut> successFunc) =>
-        result switch
+    public static IMaybe<TOut> Map<TOut>(this IMaybe result, Func<TOut> successFunc)
+    {
+        return result switch
         {
             Some => Maybe.Create(successFunc()),
             INone none => none.Cast<TOut>(),
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
-    public static async Task<IMaybe<TOut>> Map<TOut>(this IMaybe result, Func<Task<TOut>> successFunc) =>
-        result switch
+    public static async Task<IMaybe<TOut>> Map<TOut>(this IMaybe result, Func<Task<TOut>> successFunc)
+    {
+        return result switch
         {
-            Some => Maybe.Create(await successFunc()),
+            Some => Maybe.Create(await successFunc().ConfigureAwait(continueOnCapturedContext: false)),
             INone none => none.Cast<TOut>(),
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
-    public static IMaybe<TOut> Map<TIn, TOut>(this IMaybe<TIn> result, Func<TIn, TOut> mapper) =>
-        result switch
-        {
-            Some<TIn> some => Maybe.Create(mapper(some.Value)),
-            INone<TIn> none => none.Cast<TOut>(),
-            _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
-        };
+    public static IMaybe<TOut> Map<TIn, TOut>(this IMaybe<TIn> result, Func<TIn, TOut> mapper)
+    {
+        return result.Match(
+            mapper,
+            err => err.Cast<TOut>());
+    }
 
     public static async Task<IMaybe<TOut>> Map<TIn, TOut>(this IMaybe<TIn> result, Func<TIn, Task<TOut>> mapper)
     {
-        switch (result)
-        {
-            case Some<TIn> some:
-                TOut mappedResult = await mapper(some.Value).ConfigureAwait(continueOnCapturedContext: false);
-                return Maybe.Create(mappedResult);
-            case INone<TIn> none:
-                return none.Cast<TOut>();
-            default:
-                throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type");
-        }
+        return await result.Match(
+            async value =>
+            {
+                TOut mappedResult = await mapper(value).ConfigureAwait(continueOnCapturedContext: false);
+                return mappedResult;
+            },
+            err => err.Cast<TOut>());
     }
 
     public static async Task<IMaybe<TOut>> Map<TIn, TOut>(this Task<IMaybe<TIn>> result, Func<TIn, TOut> mapper)
     {
         IMaybe<TIn> awaitedResult = await result.ConfigureAwait(continueOnCapturedContext: false);
-        switch (awaitedResult)
-        {
-            case Some<TIn> some:
-                TOut mappedResult = mapper(some.Value);
-                return Maybe.Create(mappedResult);
-            case INone<TIn> none:
-                return none.Cast<TOut>();
-            default:
-                throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type");
-        }
+        return awaitedResult.Map(mapper);
     }
 
     public static async Task<IMaybe<TOut>> Map<TIn, TOut>(this Task<IMaybe<TIn>> result, Func<TIn, Task<TOut>> mapper)
     {
         IMaybe<TIn> awaitedResult = await result.ConfigureAwait(continueOnCapturedContext: false);
-        switch (awaitedResult)
-        {
-            case Some<TIn> some:
-                TOut mappedResult = await mapper(some.Value);
-                return Maybe.Create(mappedResult);
-            case INone<TIn> none:
-                return none.Cast<TOut>();
-            default:
-                throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type");
-        }
+        return await awaitedResult.Map(mapper).ConfigureAwait(continueOnCapturedContext: false);
     }
 
-    public static IMaybe FlatMap(this IMaybe result, Func<IMaybe> successFunc) =>
-        result switch
+    public static IMaybe<T> Flatten<T>(this IMaybe<IMaybe<T>> result)
+    {
+        return result switch
+        {
+            Some<IMaybe<T>> some => some.Value,
+            INone<IMaybe<T>> err => err.Cast<T>(),
+            _ => throw new ArgumentOutOfRangeException(nameof(result), "Invalid result type")
+        };
+    }
+
+    public static async Task<IMaybe<T>> Flatten<T>(this Task<IMaybe<IMaybe<T>>> result)
+    {
+        return (await result.ConfigureAwait(continueOnCapturedContext: false)).Flatten();
+    }
+
+    public static IMaybe FlatMap(this IMaybe result, Func<IMaybe> successFunc)
+    {
+        return result switch
         {
             Some => successFunc(),
             INone none => none,
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
-    public static async Task<IMaybe> FlatMap(this IMaybe result, Func<Task<IMaybe>> successFunc) =>
-        result switch
+    public static async Task<IMaybe> FlatMap(this IMaybe result, Func<Task<IMaybe>> successFunc)
+    {
+        return result switch
         {
-            Some => await successFunc(),
+            Some => await successFunc().ConfigureAwait(continueOnCapturedContext: false),
             INone none => none,
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
-    public static IMaybe<TOut> FlatMap<TOut>(this IMaybe result, Func<IMaybe<TOut>> successFunc) =>
-        result switch
+    public static IMaybe<TOut> FlatMap<TOut>(this IMaybe result, Func<IMaybe<TOut>> successFunc)
+    {
+        return result switch
         {
             Some => successFunc(),
             INone none => none.Cast<TOut>(),
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
-    public static Task<IMaybe<TOut>> FlatMap<TOut>(this IMaybe result, Func<Task<IMaybe<TOut>>> mapper) =>
-        result switch
+    public static Task<IMaybe<TOut>> FlatMap<TOut>(this IMaybe result, Func<Task<IMaybe<TOut>>> mapper)
+    {
+        return result switch
         {
             Some => mapper(),
             INone err => Task.FromResult(err.Cast<TOut>()),
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
     public static async Task<IMaybe> FlatMap(this Task<IMaybe> taskResult, Func<IMaybe> successFunc)
     {
-        IMaybe result = await taskResult;
+        IMaybe result = await taskResult.ConfigureAwait(continueOnCapturedContext: false);
         return result switch
         {
             Some => successFunc(),
@@ -115,7 +120,7 @@ public static class MaybeExtensions
 
     public static async Task<IMaybe<TOut>> FlatMap<TOut>(this Task<IMaybe> taskResult, Func<IMaybe<TOut>> successFunc)
     {
-        IMaybe result = await taskResult;
+        IMaybe result = await taskResult.ConfigureAwait(continueOnCapturedContext: false);
         return result switch
         {
             Some => successFunc(),
@@ -126,7 +131,7 @@ public static class MaybeExtensions
 
     public static async Task<IMaybe<TOut>> FlatMap<TOut>(this Task<IMaybe> taskResult, Func<Task<IMaybe<TOut>>> mapper)
     {
-        IMaybe result = await taskResult;
+        IMaybe result = await taskResult.ConfigureAwait(continueOnCapturedContext: false);
         return result switch
         {
             Some => await mapper(),
@@ -135,25 +140,18 @@ public static class MaybeExtensions
         };
     }
 
-    public static IMaybe<TOut> FlatMap<TIn, TOut>(this IMaybe<TIn> result, Func<TIn, IMaybe<TOut>> mapper) => result switch
+    public static IMaybe<TOut> FlatMap<TIn, TOut>(this IMaybe<TIn> result, Func<TIn, IMaybe<TOut>> mapper)
     {
-        Some<TIn> some => mapper(some.Value),
-        INone<TIn> none => none.Cast<TOut>(),
-        _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
-    };
+        return result
+            .Map(mapper)
+            .Flatten();
+    }
 
     public static Task<IMaybe<TOut>> FlatMap<TIn, TOut>(this IMaybe<TIn> result, Func<TIn, Task<IMaybe<TOut>>> mapper)
     {
-        switch (result)
-        {
-            case Some<TIn> some:
-                return mapper(some.Value);
-            case INone<TIn> none:
-                IMaybe<TOut> err = none.Cast<TOut>();
-                return Task.FromResult(err);
-            default:
-                throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type");
-        }
+        return result
+            .Map(mapper)
+            .Flatten();
     }
 
     public static async Task<IMaybe<TOut>> FlatMap<TIn, TOut>(
@@ -174,37 +172,43 @@ public static class MaybeExtensions
 
     public static async Task<IMaybe> FlatMap<TIn>(
         this Task<IMaybe<TIn>> result,
-        Func<TIn, Task<IMaybe>> mapper) =>
-        await result.ConfigureAwait(false) switch
+        Func<TIn, Task<IMaybe>> mapper)
+    {
+        IMaybe<IMaybe> maybe = await (await result.ConfigureAwait(continueOnCapturedContext: false))
+            .Map(mapper).ConfigureAwait(continueOnCapturedContext: false);
+        return maybe switch
         {
-            Some<TIn> some => await mapper(some.Value).ConfigureAwait(false),
+            Some<IMaybe> => Some.Instance,
             INone none => none,
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
     public static async Task<IMaybe> FlatMap<TIn>(
         this Task<IMaybe<TIn>> result,
-        Func<TIn, IMaybe> mapper) =>
-        await result.ConfigureAwait(false) switch
+        Func<TIn, IMaybe> mapper)
+    {
+        return (await result.ConfigureAwait(continueOnCapturedContext: false)).Map(mapper) switch
         {
-            Some<TIn> some => mapper(some.Value),
+            Some<IMaybe> => Some.Instance,
             INone none => none,
             _ => throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type")
         };
+    }
 
     public static void Action<TIn>(this IMaybe<TIn> result, Action<TIn> onSuccess, Action<INone>? onError = null)
     {
-        switch (result)
-        {
-            case Some<TIn> some:
-                onSuccess(some.Value);
-                break;
-            case INone<TIn> none:
-                onError?.Invoke(none);
-                break;
-            default:
-                throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type");
-        }
+        _ = result.Match(
+            value =>
+            {
+                onSuccess(value);
+                return value;
+            },
+            err =>
+            {
+                onError?.Invoke(err);
+                return err.Cast<TIn>();
+            });
     }
 
     public static async Task Action<TIn>(
@@ -213,16 +217,6 @@ public static class MaybeExtensions
         Action<INone>? onError = null)
     {
         IMaybe<TIn> awaitedResult = await result.ConfigureAwait(continueOnCapturedContext: false);
-        switch (awaitedResult)
-        {
-            case Some<TIn> some:
-                onSuccess(some.Value);
-                break;
-            case INone<TIn> none:
-                onError?.Invoke(none);
-                break;
-            default:
-                throw new ArgumentOutOfRangeException(paramName: nameof(result), message: "Invalid result type");
-        }
+        awaitedResult.Action(onSuccess, onError);
     }
 }