+++ tests/MaybeResults.UnitTests/NoneGeneratorTest.cs
@@ -71,6 +71,7 @@ public sealed class NoneGeneratorTest
                 using System.Collections.Generic;
                 using System.Linq;
                 using System.Text;
+                using System.Threading.Tasks;
                 using MaybeResults;
                 using IMaybe = MaybeResults.IMaybe;
 
@@ -278,6 +279,33 @@ public sealed class NoneGeneratorTest
                         public IMaybe ToMaybe()
                         {
                             return MyError.Create(Message, Details);
+                        }
+
+                        /// <summary>
+                        /// <![CDATA[ Executes onSome when successful and onNone when not successful. ]]>
+                        /// </summary>
+                        /// <returns>An IMaybe<TResult>.</returns>
+                        public IMaybe<TResult> Match<TResult>(Func<T, TResult> onSome, Func<INone, IMaybe<TResult>> onNone)
+                        {
+                            return onNone(this);
+                        }
+
+                        /// <summary>
+                        /// <![CDATA[ Executes onSome when successful and onNone when not successful. ]]>
+                        /// </summary>
+                        /// <returns>An IMaybe</returns>
+                        public IMaybe Match<TResult>(Func<T, IMaybe> onSome, Func<INone, IMaybe> onNone)
+                        {
+                            return onNone(this);
+                        }
+
+                        /// <summary>
+                        /// <![CDATA[ Executes onSome when successful and onNone when not successful. ]]>
+                        /// </summary>
+                        /// <returns>A Task<IMaybe<TResult>>.</returns>
+                        public Task<IMaybe<TResult>> Match<TResult>(Func<T, Task<TResult>> onSome, Func<INone, IMaybe<TResult>> onNone)
+                        {
+                            return Task.FromResult(onNone(this));
                         }
                     }
                 }