+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [2.0.2]
+- Added Match method to base IMaybe interface and use it in order to fix invariance issue.
+
+## [2.0.1]
+- Added two missing FlatMap Overloads.
+
 ## [2.0.0]
 ### Added
 - Support for Linq syntax.